Pass validation of all tests
